# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v3/contents/generations/tasks": {
      "post": {
        "summary": "创建 Seedance V3 视频任务",
        "deprecated": false,
        "description": "统一创建 Seedance 2.0 异步视频任务。支持 dreamina-seedance-2-0-hc、doubao-seedance-2-0-filter-off、doubao-seedance-2-0 和 doubao-seedance-2-0-fast。响应使用本站任务 ID，不暴露上游 ID。HC 模型会自动上传公网图片；Doubao 模型仅在媒体项 contains_face=true 时自动创建素材、等待 Active 并替换为 asset:// 引用。",
        "operationId": "createSeedanceV3VideoTask",
        "tags": [
          "视频生成/Seedance V3兼容格式"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedanceV3VideoRequest"
              },
              "examples": {
                "text_to_video": {
                  "summary": "文生视频",
                  "value": {
                    "model": "dreamina-seedance-2-0-hc",
                    "content": [
                      {
                        "type": "text",
                        "text": "一只猫在阳光明媚的草地上奔跑"
                      }
                    ],
                    "resolution": "720p",
                    "ratio": "16:9",
                    "duration": 5,
                    "watermark": false
                  }
                },
                "image_to_video": {
                  "summary": "图生视频",
                  "value": {
                    "model": "dreamina-seedance-2-0-hc",
                    "content": [
                      {
                        "type": "image_url",
                        "image_url": {
                          "url": "https://example.com/input.jpg"
                        },
                        "role": "reference_image"
                      },
                      {
                        "type": "text",
                        "text": "让图中的猫动起来，在草地上奔跑"
                      }
                    ],
                    "resolution": "720p",
                    "ratio": "16:9",
                    "duration": 5,
                    "watermark": false
                  }
                },
                "video_to_video": {
                  "summary": "视频生视频",
                  "value": {
                    "model": "dreamina-seedance-2-0-hc",
                    "content": [
                      {
                        "type": "video_url",
                        "video_url": {
                          "url": "https://example.com/input.mp4"
                        },
                        "role": "reference_video"
                      },
                      {
                        "type": "text",
                        "text": "延续这段视频并增加更多动感"
                      }
                    ],
                    "resolution": "720p",
                    "ratio": "16:9",
                    "duration": 5,
                    "watermark": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建视频任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedanceV3VideoTask"
                },
                "example": {
                  "id": "task_2b7f10d34b6449ada4e53d76e31a5290",
                  "model": "dreamina-seedance-2-0-hc",
                  "status": "queued",
                  "content": {},
                  "created_at": 1784217600
                }
              }
            }
          },
          "400": {
            "description": "请求参数或 content 结构错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API Key 缺失或无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "请求频率过高",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误、上游素材上传失败或任务创建失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/videos": {
      "post": {
        "summary": "创建视频 ",
        "deprecated": false,
        "description": "OpenAI 兼容的视频生成接口。\n\n参考文档: https://platform.openai.com/docs/api-reference/videos/create\n",
        "operationId": "createVideo",
        "tags": [
          "视频生成/Sora兼容格式"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "description": "模型名称",
                    "example": "sora-2",
                    "type": "string"
                  },
                  "prompt": {
                    "description": "提示词",
                    "example": "cute cat dance",
                    "type": "string"
                  },
                  "seconds": {
                    "description": "生成秒数",
                    "example": "8",
                    "type": "string"
                  },
                  "input_reference": {
                    "format": "binary",
                    "type": "string",
                    "description": "参考图片文件",
                    "example": ""
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建视频任务",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "视频 ID"
                    },
                    "object": {
                      "type": "string",
                      "description": "对象类型"
                    },
                    "model": {
                      "type": "string",
                      "description": "使用的模型"
                    },
                    "status": {
                      "type": "string",
                      "description": "任务状态"
                    },
                    "progress": {
                      "type": "integer",
                      "description": "进度百分比"
                    },
                    "created_at": {
                      "type": "integer",
                      "description": "创建时间戳"
                    },
                    "seconds": {
                      "type": "string",
                      "description": "视频时长"
                    },
                    "completed_at": {
                      "type": "integer",
                      "description": "完成时间戳"
                    },
                    "expires_at": {
                      "type": "integer",
                      "description": "过期时间戳"
                    },
                    "size": {
                      "type": "string",
                      "description": "视频尺寸"
                    },
                    "error": {
                      "$ref": "#/components/schemas/OpenAIVideoError"
                    },
                    "metadata": {
                      "type": "object",
                      "description": "额外元数据",
                      "additionalProperties": true,
                      "properties": {}
                    }
                  },
                  "required": [
                    "id",
                    "object",
                    "model",
                    "status",
                    "progress",
                    "created_at",
                    "seconds"
                  ]
                },
                "example": {
                  "id": "sora-2-123456",
                  "object": "video",
                  "model": "sora-2",
                  "status": "queued",
                  "progress": 0,
                  "created_at": 1764347090922,
                  "seconds": "8"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SeedanceV3VideoRequest": {
        "type": "object",
        "description": "Seedance V3 兼容视频生成请求",
        "properties": {
          "model": {
            "type": "string",
            "description": "视频生成模型",
            "enum": [
              "dreamina-seedance-2-0-hc",
              "doubao-seedance-2-0-filter-off",
              "doubao-seedance-2-0",
              "doubao-seedance-2-0-fast"
            ],
            "example": "dreamina-seedance-2-0-hc"
          },
          "content": {
            "type": "array",
            "description": "输入内容数组，支持文本、图片和视频",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/SeedanceV3ContentItem"
            }
          },
          "resolution": {
            "type": "string",
            "description": "输出分辨率，默认 480p",
            "enum": [
              "480p",
              "720p",
              "1080p",
              "4K"
            ],
            "default": "480p"
          },
          "ratio": {
            "type": "string",
            "description": "输出画面比例，默认 16:9",
            "enum": [
              "21:9",
              "16:9",
              "9:16",
              "3:4",
              "1:1",
              "4:3",
              "adaptive"
            ],
            "default": "16:9"
          },
          "duration": {
            "description": "视频时长。HC 支持 4-15 秒；Doubao 还支持 -1 由模型决定",
            "oneOf": [
              {
                "type": "integer",
                "minimum": 4,
                "maximum": 15
              },
              {
                "type": "integer",
                "enum": [
                  -1
                ]
              }
            ],
            "default": 5,
            "example": 5
          },
          "watermark": {
            "type": "boolean",
            "description": "是否添加水印",
            "default": false
          },
          "generate_audio": {
            "type": "boolean",
            "description": "是否生成音频",
            "default": false
          },
          "tools": {
            "type": "array",
            "description": "Doubao 工具列表",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "web_search"
                }
              },
              "required": [
                "type"
              ]
            }
          },
          "service_tier": {
            "type": "string",
            "description": "Doubao 服务等级，透传上游"
          },
          "draft": {
            "type": "boolean",
            "description": "Doubao 草稿模式，透传上游"
          },
          "frames": {
            "type": "integer",
            "description": "Doubao 输出帧数，透传上游"
          },
          "camera_fixed": {
            "type": "boolean",
            "description": "Doubao 固定镜头参数，透传上游"
          }
        },
        "required": [
          "model",
          "content"
        ]
      },
      "SeedanceV3VideoTask": {
        "type": "object",
        "description": "Seedance V3 视频任务",
        "properties": {
          "id": {
            "type": "string",
            "description": "本站生成的公开任务 ID",
            "example": "task_2b7f10d34b6449ada4e53d76e31a5290"
          },
          "model": {
            "type": "string",
            "description": "视频生成模型",
            "example": "dreamina-seedance-2-0-hc"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "expired"
            ]
          },
          "content": {
            "type": "object",
            "description": "生成结果，任务成功后包含 video_url",
            "properties": {
              "video_url": {
                "type": "string",
                "description": "具有时效性的视频下载地址",
                "example": "https://example.com/generated-video.mp4"
              }
            }
          },
          "duration_seconds": {
            "type": "integer",
            "minimum": 0,
            "description": "生成视频时长（秒）"
          },
          "outputs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "成功后的视频下载地址列表"
          },
          "usage": {
            "type": "object",
            "description": "上游返回的用量信息；任务未完成或上游未返回用量时可能省略",
            "properties": {
              "completion_tokens": {
                "type": "integer",
                "minimum": 0,
                "description": "视频生成消耗的输出 Token"
              },
              "total_tokens": {
                "type": "integer",
                "minimum": 0,
                "description": "本次任务消耗的总 Token"
              }
            }
          },
          "error": {
            "type": "object",
            "nullable": true,
            "description": "任务失败信息",
            "properties": {
              "code": {
                "type": "string",
                "description": "错误码"
              },
              "message": {
                "type": "string",
                "description": "错误信息"
              }
            }
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "任务创建时间，Unix 秒级时间戳"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "任务最近更新时间，Unix 秒级时间戳"
          },
          "completed_at": {
            "type": "integer",
            "format": "int64",
            "description": "任务完成时间，Unix 秒级时间戳"
          },
          "last_frame_url": {
            "type": "string",
            "description": "视频最后一帧地址（可选）"
          }
        },
        "required": [
          "id",
          "model",
          "status"
        ]
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}